Refactored routes. Added data associations.
