+++ models/review.js
@@ -4,7 +4,13 @@ const mongoose = require('mongoose');
 // SCHEMA SET UP
 const reviewSchema = new mongoose.Schema({
 	text: String,
-	author: String
+	author: {
+		id: {
+			type: mongoose.Schema.Types.ObjectId,
+			ref: 'User'  //name of model we are referencing
+		},
+		username: String
+	}
 });
 
 // COMPILE SCHEMA INTO MODEL & EXPORT
