+++ views/bookstores/show.ejs
@@ -6,6 +6,7 @@
 		<div class='card-body'>
 			<h1 id='bookstore-show-name' class='card-title'><%= bookstore.name %></h1>
 			<p class='card-text'><%= bookstore.description %></p>
+			<p class='card-text text-info-edit'>Submitted by: <%= bookstore.author.username %></p>
 			<a class='btn btn-info btn-sm' href='/bookstores/<%= bookstore._id %>/reviews/new'>Add a Review</a>
 			<a class='btn btn-outline-info btn-sm' href='/bookstores'>Go Back</a>
 
@@ -14,7 +15,7 @@
 			<h2 id='bookstore-show-sub' class='card-subtitle mb-3'>Reviews</h2>
 			<% bookstore.reviews.forEach(function(review){ %>
 				<p class='card-text'>
-					<span class='card-subtitle h6'><%= review.author %>:</span>
+					<span class='card-subtitle h6'><%= review.author.username %>:</span>
 					<%= review.text %>
 				</p>
 			<% }) %>
