+++ views/reviews/new.ejs
@@ -7,12 +7,7 @@
 	<div class='w-50 mx-auto'> <!--50% width, center horizontally-->
 		<form action='/bookstores/<%= bookstore._id %>/reviews' method='POST'> <!--form sends data via post request-->
 		  <div class='form-group'>
-		    <label for='author'>Your Name</label>
-		    <input type='text' class='form-control' name='review[author]' id='author' placeholder='Enter your first name'>
-		  </div>
-		  <div class='form-group'>
-		    <label for='text'>Review</label>
-		    <textarea class='form-control' name='review[text]' id='text' rows='5' placeholder='Enter your review'></textarea>
+		    <textarea class='form-control' name='review[text]' rows='5' placeholder='Enter your review here'></textarea>
 		  </div>
 		  <button type='submit' class='btn btn-info'>Submit</button>
 		  <a class='btn btn-outline-info' href='/bookstores/<%= bookstore._id %>'>Go Back</a>
